Extract SignupSiteTitle to a separate component.
