+++ assets/stylesheets/_components.scss
@@ -112,6 +112,7 @@
 @import 'blocks/upgrade-nudge-expanded/style';
 @import 'components/seo/preview-upgrade-nudge/style';
 @import 'components/signup-form/style';
+@import 'components/signup-site-title/style';
 @import 'components/site-icon/style';
 @import 'components/site-selector/style';
 @import 'components/site-selector-modal/style';
@@ -406,7 +407,6 @@
 @import 'signup/steps/plans/style';
 @import 'signup/steps/paid-plans-only/style';
 @import 'signup/steps/site/style';
-@import 'signup/steps/site-title/style';
 @import 'signup/validation-fieldset/style';
 @import 'support/support-user/style';
 @import 'vip/vip-logs/style';
